==> src/05_github-networks/01_full-user-networks/02_full-ctr-nets-cum-analysis/00_install_packages_for_slurms.R <==
################################################################################## install packages (for slurm)

install.packages("igraph", repos = "http://cran.us.r-project.org")
install.packages("lubridate", repos = "http://cran.us.r-project.org")
